sql test cases. move code from mod to runner.rs
